+++ view/forum/listTopics.php
@@ -14,26 +14,31 @@ $catId=$categorie->getId();
 <?php } ?>
 
 <?php
-foreach($topics as $topic){
-    echo "<p><a href='index.php?ctrl=forum&action=listPosts&id=".$topic->getId()."'>".$topic->getTitreTopic()."</a>";
-    if ($topic->getVerrouTopic()){
-        echo " [VERROUILLÉ]";
-    } 
-    if(\App\Session::getUser()){
-        if (\App\Session::getUser()->getId() == $topic->getUser()->getId()){
-            echo " [<a href='index.php?ctrl=forum&action=editTopicForm&id=".$topic->getId()."'>Éditer</a> /
-            <a href='index.php?ctrl=forum&action=deleteTopic&id=".$topic->getId()."'>Supprimer</a> / ";
-            
-            if($topic->getVerrouTopic()){
-                echo "<a href='index.php?ctrl=forum&action=unlockTopic&id=".$topic->getId()."'>Déverrouiller</a>";
-            }
-            else{
-                echo "<a href='index.php?ctrl=forum&action=lockTopic&id=".$topic->getId()."'>Verrouiller</a>";
-            }
-            echo "]";
+if($topics){
+    foreach($topics as $topic){
+        echo "<p><a href='index.php?ctrl=forum&action=listPosts&id=".$topic->getId()."'>".$topic->getTitreTopic()."</a>";
+        if ($topic->getVerrouTopic()){
+            echo " [VERROUILLÉ]";
         } 
-    }   
-    echo "<br>Date ".$topic->getDateCreaTopic()." - Auteur :  ".$topic->getUser()."</p>";
+        if(\App\Session::getUser()){
+            if (\App\Session::getUser()->getId() == $topic->getUser()->getId()){
+                echo " [<a href='index.php?ctrl=forum&action=editTopicForm&id=".$topic->getId()."'>Éditer</a> /
+                <a href='index.php?ctrl=forum&action=deleteTopic&id=".$topic->getId()."'>Supprimer</a> / ";
+                
+                if($topic->getVerrouTopic()){
+                    echo "<a href='index.php?ctrl=forum&action=unlockTopic&id=".$topic->getId()."'>Déverrouiller</a>";
+                }
+                else{
+                    echo "<a href='index.php?ctrl=forum&action=lockTopic&id=".$topic->getId()."'>Verrouiller</a>";
+                }
+                echo "]";
+            } 
+        }   
+        echo "<br>Date ".$topic->getDateCreaTopic()." - Auteur :  ".$topic->getUser()."</p>";
+    }
+}
+else{
+    echo "<p>Aucun topic dans cette catégorie</p>";
 }
 
 ?>
